decouple from CLI part2
