+++ cli/main.rs
@@ -55,15 +55,22 @@ use crate::file_fetcher::FileFetcher;
 use crate::file_watcher::ModuleResolutionResult;
 use crate::flags::DenoSubcommand;
 use crate::flags::Flags;
+use crate::fmt_errors::PrettyJsError;
 use crate::import_map::ImportMap;
 use crate::media_type::MediaType;
+use crate::module_loader::CliModuleLoader;
+use crate::ops::worker_host::CreateWebWorkerCb;
 use crate::permissions::Permissions;
 use crate::program_state::exit_unstable;
 use crate::program_state::ProgramState;
+use crate::source_maps::apply_source_map;
 use crate::specifier_handler::FetchHandler;
 use crate::standalone::create_standalone_binary;
 use crate::tools::installer::infer_name_from_url;
-use crate::worker::create_main_worker;
+use crate::web_worker::WebWorker;
+use crate::web_worker::WebWorkerOptions;
+use crate::worker::MainWorker;
+use crate::worker::WorkerOptions;
 use deno_core::error::generic_error;
 use deno_core::error::AnyError;
 use deno_core::futures::future::FutureExt;
@@ -85,6 +92,137 @@ use std::path::PathBuf;
 use std::pin::Pin;
 use std::rc::Rc;
 use std::sync::Arc;
+
+fn create_web_worker_callback(
+  program_state: Arc<ProgramState>,
+) -> Arc<CreateWebWorkerCb> {
+  Arc::new(
+    move |name, worker_id, permissions, main_module, has_deno_namespace| {
+      let global_state_ = program_state.clone();
+      let js_error_create_fn = Rc::new(move |core_js_error| {
+        let source_mapped_error =
+          apply_source_map(&core_js_error, global_state_.clone());
+        PrettyJsError::create(source_mapped_error)
+      });
+
+      let attach_inspector = program_state.maybe_inspector_server.is_some()
+        || program_state.flags.coverage;
+      let maybe_inspector_server = program_state.maybe_inspector_server.clone();
+
+      let module_loader =
+        CliModuleLoader::new_for_worker(program_state.clone());
+      let create_web_worker_cb =
+        create_web_worker_callback(program_state.clone());
+
+      let options = WebWorkerOptions {
+        args: program_state.flags.argv.clone(),
+        apply_source_maps: true,
+        debug_flag: program_state
+          .flags
+          .log_level
+          .map_or(false, |l| l == log::Level::Debug),
+        unstable: program_state.flags.unstable,
+        ca_filepath: program_state.flags.ca_file.clone(),
+        seed: program_state.flags.seed,
+        module_loader,
+        create_web_worker_cb,
+        js_error_create_fn: Some(js_error_create_fn),
+        has_deno_namespace,
+        attach_inspector,
+        maybe_inspector_server,
+      };
+
+      let mut worker = WebWorker::from_options(
+        name,
+        permissions,
+        main_module,
+        worker_id,
+        &options,
+      );
+
+      // NOTE(bartlomieju): ProgramState is CLI only construct,
+      // hence we're not using it in `Self::from_options`.
+      {
+        let js_runtime = &mut worker.js_runtime;
+        js_runtime
+          .op_state()
+          .borrow_mut()
+          .put::<Arc<ProgramState>>(program_state.clone());
+        // Applies source maps - works in conjuction with `js_error_create_fn`
+        // above
+        ops::errors::init(js_runtime);
+        if has_deno_namespace {
+          ops::runtime_compiler::init(js_runtime);
+        }
+      }
+      worker.bootstrap(&options);
+
+      worker
+    },
+  )
+}
+
+pub fn create_main_worker(
+  program_state: &Arc<ProgramState>,
+  main_module: ModuleSpecifier,
+  permissions: Permissions,
+) -> MainWorker {
+  let module_loader = CliModuleLoader::new(program_state.clone());
+
+  let global_state_ = program_state.clone();
+
+  let js_error_create_fn = Rc::new(move |core_js_error| {
+    let source_mapped_error =
+      apply_source_map(&core_js_error, global_state_.clone());
+    PrettyJsError::create(source_mapped_error)
+  });
+
+  let attach_inspector = program_state.maybe_inspector_server.is_some()
+    || program_state.flags.repl
+    || program_state.flags.coverage;
+  let maybe_inspector_server = program_state.maybe_inspector_server.clone();
+  let should_break_on_first_statement =
+    program_state.flags.inspect_brk.is_some();
+
+  let create_web_worker_cb = create_web_worker_callback(program_state.clone());
+
+  let options = WorkerOptions {
+    apply_source_maps: true,
+    args: program_state.flags.argv.clone(),
+    debug_flag: program_state
+      .flags
+      .log_level
+      .map_or(false, |l| l == log::Level::Debug),
+    unstable: program_state.flags.unstable,
+    ca_filepath: program_state.flags.ca_file.clone(),
+    seed: program_state.flags.seed,
+    js_error_create_fn: Some(js_error_create_fn),
+    create_web_worker_cb,
+    attach_inspector,
+    maybe_inspector_server,
+    should_break_on_first_statement,
+    module_loader,
+  };
+
+  let mut worker = MainWorker::from_options(main_module, permissions, &options);
+
+  // NOTE(bartlomieju): ProgramState is CLI only construct,
+  // hence we're not using it in `Self::from_options`.
+  {
+    let js_runtime = &mut worker.js_runtime;
+    js_runtime
+      .op_state()
+      .borrow_mut()
+      .put::<Arc<ProgramState>>(program_state.clone());
+    // Applies source maps - works in conjuction with `js_error_create_fn`
+    // above
+    ops::errors::init(js_runtime);
+    ops::runtime_compiler::init(js_runtime);
+    worker.bootstrap(&options);
+  }
+
+  worker
+}
 
 fn write_to_stdout_ignore_sigpipe(bytes: &[u8]) -> Result<(), std::io::Error> {
   use std::io::ErrorKind;
